The EshyUI library is complete to a reasonable degree. The next important thing is to allow creation of euiImageEntities with image data, and not just a png file. Additionally, the renderer needs to be made into a batch renderer.

+++ EshyUI/EshyUI/Source/EshyUI.cpp
@@ -14,7 +14,7 @@ GLFWwindow* InitEshyUI(const std::string& WindowName, float Width, float Height)
 	glfwWindowHint(GLFW_CONTEXT_VERSION_MINOR, 6);
 	glfwWindowHint(GLFW_OPENGL_PROFILE, GLFW_OPENGL_CORE_PROFILE);
 
-	GLFWwindow* window = glfwCreateWindow(Width, Height, WindowName.c_str(), NULL, NULL);
+	GLFWwindow* window = glfwCreateWindow((int)Width, (int)Height, WindowName.c_str(), NULL, NULL);
 	eshyui_glfw_check(window);
 	glfwMakeContextCurrent(window);
 	glfwSwapInterval(1);
